requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.29.0  # Driver PostgreSQL asíncrono
aiosqlite==0.20.0  # Driver SQLite asíncrono
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
fastapi-cors==0.0.6
structlog==24.4.0
